+++ src/main/java/com/zheng/aicommunitybackend/service/PostCommentsService.java
@@ -57,14 +57,14 @@ public interface PostCommentsService extends IService<PostComments> {
      * @param postId 帖子ID
      * @return 评论总数
      */
-    Integer countCommentsByPostId(Long postId);
+    Long countCommentsByPostId(Long postId);
     
     /**
      * 批量获取帖子评论总数
      * @param postIds 帖子ID集合
      * @return 帖子ID与评论数量的映射
      */
-    Map<Long, Integer> batchCountCommentsByPostIds(List<Long> postIds);
+    Map<Long, Long> batchCountCommentsByPostIds(List<Long> postIds);
     
     /**
      * 管理端分页查询评论
